added to show specific reservation
showing reservation by confirmation code so the user can edit

+++ Assignment 2/bistro_prototype/bistro_prototype_server/src/entities/Reservation.java
@@ -12,7 +12,7 @@ public class Reservation implements Serializable{
 	private int subscriberId;
 	private LocalDate dateOfPlacingOrder;
 	
-	public Reservation(long reservationID,LocalDate reservationDate,int numberOfGuests,int confirmationCode,int subscriberId,LocalDate dateOfPlacingOrder) {
+	public Reservation(int reservationID,LocalDate reservationDate,int numberOfGuests,int confirmationCode,int subscriberId,LocalDate dateOfPlacingOrder) {
 		this.reservationID = reservationID;
 		this.reservationDate = reservationDate;
 		this.numberOfGuests = numberOfGuests;
